Fix volume data type handling in Finnhub candle response

+++ app/src/main/java/com/vienna/app/data/remote/dto/FinnhubDto.kt
@@ -68,7 +68,7 @@ data class FinnhubCandleResponse(
     @SerialName("t")
     val timestamps: List<Long>? = null,
     @SerialName("v")
-    val volumes: List<Long>? = null,
+    val volumes: List<Double>? = null,
     @SerialName("s")
     val status: String = "ok"
 )
